backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-CORS>=4.0.0
Flask-Migrate>=4.0.5
psycopg2-binary>=2.9.9
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
